refactor(ui): split section wrappers into per-file parts with namespace exports
Move each composed section wrapper into its own file and replace
Object.assign compound components with `export * as` namespace
exports for tree-shakability.
