+++ packages/ui/src/composed/OrganizationProfile/GeneralDeleteOrganization.tsx
@@ -1,0 +1,9 @@
+import type { ReactNode } from 'react';
+
+import { OrganizationDeleteSection } from '../../components/OrganizationProfile/OrganizationGeneralPage';
+import { useRequirePage } from '../useRequirePage';
+
+export function GeneralDeleteOrganization(): ReactNode {
+  if (!useRequirePage('GeneralDeleteOrganization')) return null;
+  return <OrganizationDeleteSection />;
+}
