+++ packages/ui/src/composed/OrganizationProfile/parts.ts
@@ -1,0 +1,10 @@
+export { OrganizationProfileProvider as Root } from './OrganizationProfileProvider';
+export { General } from './General';
+export { Members } from './Members';
+export { Billing } from './Billing';
+export { APIKeys } from './APIKeys';
+export { ConfigureSSO } from './ConfigureSSO';
+export { GeneralOrganizationProfile } from './GeneralOrganizationProfile';
+export { GeneralVerifiedDomains } from './GeneralVerifiedDomains';
+export { GeneralLeaveOrganization } from './GeneralLeaveOrganization';
+export { GeneralDeleteOrganization } from './GeneralDeleteOrganization';
