+++ packages/ui/src/composed/UserProfile/AccountConnectedAccounts.tsx
@@ -1,0 +1,9 @@
+import type { ReactNode } from 'react';
+
+import { AccountConnectedAccounts as AccountConnectedAccountsSection } from '../../components/UserProfile/AccountSections';
+import { useRequirePage } from '../useRequirePage';
+
+export function AccountConnectedAccounts(): ReactNode {
+  if (!useRequirePage('AccountConnectedAccounts')) return null;
+  return <AccountConnectedAccountsSection />;
+}
