+++ packages/ui/src/composed/UserProfile/AccountEmails.tsx
@@ -1,0 +1,9 @@
+import type { ReactNode } from 'react';
+
+import { AccountEmails as AccountEmailsSection } from '../../components/UserProfile/AccountSections';
+import { useRequirePage } from '../useRequirePage';
+
+export function AccountEmails(): ReactNode {
+  if (!useRequirePage('AccountEmails')) return null;
+  return <AccountEmailsSection />;
+}
