+++ packages/ui/src/composed/UserProfile/SecurityActiveDevices.tsx
@@ -1,0 +1,9 @@
+import type { ReactNode } from 'react';
+
+import { ActiveDevicesSection } from '../../components/UserProfile/ActiveDevicesSection';
+import { useRequirePage } from '../useRequirePage';
+
+export function SecurityActiveDevices(): ReactNode {
+  if (!useRequirePage('SecurityActiveDevices')) return null;
+  return <ActiveDevicesSection />;
+}
