+++ packages/ui/src/composed/UserProfile/SecurityDelete.tsx
@@ -1,0 +1,9 @@
+import type { ReactNode } from 'react';
+
+import { SecurityDelete as SecurityDeleteSection } from '../../components/UserProfile/SecuritySections';
+import { useRequirePage } from '../useRequirePage';
+
+export function SecurityDelete(): ReactNode {
+  if (!useRequirePage('SecurityDelete')) return null;
+  return <SecurityDeleteSection />;
+}
