+++ packages/ui/src/composed/UserProfile/SecurityMfa.tsx
@@ -1,0 +1,9 @@
+import type { ReactNode } from 'react';
+
+import { SecurityMfa as SecurityMfaSection } from '../../components/UserProfile/SecuritySections';
+import { useRequirePage } from '../useRequirePage';
+
+export function SecurityMfa(): ReactNode {
+  if (!useRequirePage('SecurityMfa')) return null;
+  return <SecurityMfaSection />;
+}
