+++ packages/ui/src/composed/UserProfile/SecurityPasskeys.tsx
@@ -1,0 +1,9 @@
+import type { ReactNode } from 'react';
+
+import { SecurityPasskeys as SecurityPasskeysSection } from '../../components/UserProfile/SecuritySections';
+import { useRequirePage } from '../useRequirePage';
+
+export function SecurityPasskeys(): ReactNode {
+  if (!useRequirePage('SecurityPasskeys')) return null;
+  return <SecurityPasskeysSection />;
+}
