+++ packages/ui/src/composed/UserProfile/SecurityPassword.tsx
@@ -1,0 +1,9 @@
+import type { ReactNode } from 'react';
+
+import { SecurityPassword as SecurityPasswordSection } from '../../components/UserProfile/SecuritySections';
+import { useRequirePage } from '../useRequirePage';
+
+export function SecurityPassword(): ReactNode {
+  if (!useRequirePage('SecurityPassword')) return null;
+  return <SecurityPasswordSection />;
+}
