+++ packages/ui/src/composed/UserProfile/parts.ts
@@ -1,0 +1,17 @@
+export { UserProfileProvider as Root } from './UserProfileProvider';
+export { Account } from './Account';
+export { Security } from './Security';
+export { Billing } from './Billing';
+export { APIKeys } from './APIKeys';
+export { AccountProfile } from './AccountProfile';
+export { AccountUsername } from './AccountUsername';
+export { AccountEmails } from './AccountEmails';
+export { AccountPhone } from './AccountPhone';
+export { AccountConnectedAccounts } from './AccountConnectedAccounts';
+export { AccountEnterpriseAccounts } from './AccountEnterpriseAccounts';
+export { AccountWeb3 } from './AccountWeb3';
+export { SecurityPassword } from './SecurityPassword';
+export { SecurityPasskeys } from './SecurityPasskeys';
+export { SecurityMfa } from './SecurityMfa';
+export { SecurityActiveDevices } from './SecurityActiveDevices';
+export { SecurityDelete } from './SecurityDelete';
